cleanup code

+++ main/localization/position_sender.cc
@@ -4,7 +4,7 @@
 
 PositionSender::PositionSender(std::vector<int> tag_ids)
     : tag_ids_(tag_ids), instance_(nt::NetworkTableInstance::GetDefault()) {
-  for (int i = 0; i < tag_ids_.size(); i++) {
+  for (size_t i = 0; i < tag_ids_.size(); i++) {
     std::shared_ptr<nt::NetworkTable> table =
         instance_.GetTable("orin/pose_estimate/" + std::to_string(tag_ids_[i]));
 
@@ -35,8 +35,8 @@ PositionSender::PositionSender(std::vector<int> tag_ids)
 
 void PositionSender::Send(
     std::vector<PoseEstimator::position_estimate_t> position_estimates) {
-  for (int i = 0; i < tag_ids_.size(); i++) {
-    for (int j = 0; j < position_estimates.size(); j++) {
+  for (size_t i = 0; i < tag_ids_.size(); i++) {
+    for (size_t j = 0; j < position_estimates.size(); j++) {
       status_[i].Set(false);
       if (tag_ids_[i] == position_estimates[j].tag_id) {
         translation_x_publisher_[i].Set(position_estimates[j].translation.y);
